P11_Categorisez_automatiquement_des_questions_aymeric_bailleul/_projet/pyproject.toml
[project]
name = "p11-rag-env"
version = "0.1.0"
description = "Developpement d'un assistant pour la recommandation d'evenements culturels"
authors = [
    {name = "Aymeric Bailleul"}
]
readme = "README.md"
requires-python = ">=3.11.0"
dependencies = [
    "langchain>=0.3.19",
    "langchain-community>=0.3.19",
    "langchain-mistralai>=0.2.7",
    "faiss-cpu>=1.13.2",
    "mistralai>=1.12.0",
    "pandas>=2.3.0",
    "pyarrow>=23.0.0",
    "requests>=2.32.5",
    "python-dotenv>=1.2.1",
    "numpy>=2.4.2",
    "tiktoken>=0.12.0",
    "streamlit>=1.54.0",
    "ragas>=0.4.3",
    "datasets>=3.3.0",
    "nest-asyncio>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "notebook>=7.5.3",
    "ipykernel>=7.1.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
